Se agrega los item para la categoria de las peliculas y el casting de las peliculas.

+++ app/src/main/java/com/code93/e_movie/domain/model/DetailsModel.kt
@@ -12,6 +12,6 @@ data class DetailsModel(
     val status: String,
     val tagline: String,
     val title: String,
-    val voteAverage: Int,
-    val voteCount: Int
+    val voteAverage: Double,
+    val voteCount: Double
 )
